+++ test/Feature/CleanRegex/Replace/only/ReplacePatternTest.php
@@ -3,7 +3,7 @@ namespace Test\Feature\TRegx\CleanRegex\Replace\only;
 
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
-use TRegx\CleanRegex\Match\Details\ReplaceDetail;
+use TRegx\CleanRegex\Replace\Details\ReplaceDetail;
 
 /**
  * @coversNothing