+++ test/Unit/CleanRegex/Match/Details/ReplaceDetailTest.php
@@ -11,10 +11,10 @@ use TRegx\CleanRegex\Match\Details\Detail;
 use TRegx\CleanRegex\Match\Details\DuplicateName;
 use TRegx\CleanRegex\Match\Details\Groups\IndexedGroups;
 use TRegx\CleanRegex\Match\Details\Groups\NamedGroups;
-use TRegx\CleanRegex\Match\Details\ReplaceDetail;
+use TRegx\CleanRegex\Replace\Details\ReplaceDetail;
 
 /**
- * @covers \TRegx\CleanRegex\Match\Details\ReplaceDetail
+ * @covers \TRegx\CleanRegex\Replace\Details\ReplaceDetail
  */
 class ReplaceDetailTest extends TestCase
 {